+++ src/main/kotlin/me/bors/slack/share/ui/SlackShareFileAction.kt
@@ -1,9 +1,11 @@
-package me.bors.slack.share
+package me.bors.slack.share.ui
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
 import com.intellij.openapi.vfs.VirtualFile
+import me.bors.slack.share.SlackClient
+import me.bors.slack.share.SlackConversationsProcessor
 
 class SlackShareFileAction : AnAction() {
     override fun actionPerformed(e: AnActionEvent) {
